requirements.txt
Flask==3.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.22.0
werkzeug==3.0.1
